# gateway/openapi.yaml
openapi: 3.0.3
info:
  title: Neo-Bank Gateway API
  version: 0.1.0
  description: >
    HTTP contract exposed by the Gateway (services/gateway) to browser
    clients. /auth/* proxies to auth-svc, /accounts/* proxies to
    accounts-svc, /transfers/* and /deposits/* proxy to transfers-svc —
    see gateway/proxy.go and gateway/main.go's newHandler for the routing
    table and gateway/middleware.go for which paths require a bearer
    token.

    Only endpoints the frontend actually consumes are documented here, not
    every route accounts-svc happens to expose behind the Gateway (e.g.
    GET /accounts/{id} and PATCH /accounts/{id}/status are internal/ops
    surface, not used by this client, and intentionally left out). For the
    same reason, POST /webhooks/stripe is deliberately absent: Stripe
    calls it, never the browser, so it isn't part of this client contract
    — see services/transfers-svc/webhook.go for that endpoint instead.
servers:
  - url: http://localhost:8080
    description: >
      Gateway, local dev. The frontend dev server never talks here
      directly — it proxies /api/* to this origin, stripping /api first
      (see frontend/vite.config.ts).
security: []
paths:
...
  /transfers:
    post:
      operationId: createTransfer
      summary: >
        Create a transfer from the caller's own account to a recipient
        identified by IBAN. sender_account_id is never taken from the
        request body — it's derived from the bearer token. Only transfers
        within this bank are supported (no SEPA/inter-bank rail) — a
        structurally valid IBAN for a different institution is rejected
        with 400, distinctly from a malformed IBAN or an unknown recipient
        within this bank.
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: >
            Client-generated UUID for this transfer attempt. Retrying the
            exact same attempt (network failure, no response received) must
            reuse the same key; a genuinely new transfer needs a new one.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransferRequest'
      responses:
        '201':
          description: A definite outcome was reached (completed, failed, or rejected).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResult'
        '200':
          description: >
            A transfer with this Idempotency-Key already existed — its
            current state is returned as-is, without moving money again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResult'
        '202':
          description: >
            The outcome is uncertain and the transfer is left pending. Two
            distinct causes share this same shape (TransferResult with
            status "pending" and a message) with no separate discriminator
            field — callers must match on message text: ledger-svc did not
            respond definitively ("transfer status unknown, still
            processing"), or fraud-svc was unreachable and the check fails
            closed ("fraud check unavailable, transfer still pending").
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Recipient not found, or the caller has no account yet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Recipient account is closed, or sender account is not active.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Idempotency-Key was already used with different parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: >
            Too many recipient-resolve attempts for this caller recently —
            see accounts-svc's per-user rate limit on resolving an IBAN
            (README, "Honest limitations"). Distinct from a generic
            TooManyRequests: this specific limit exists because an
            endpoint that answers "this account exists"/"it doesn't" is an
            enumeration oracle, and a checksum-valid IBAN narrows the space
            worth guessing rather than making it safe to guess.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      operationId: listTransfers
      summary: The caller's own transfer history (sent and received), newest first.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          description: >
            Opaque continuation token from a previous response's
            next_cursor. Omit to fetch the first page.
          schema:
            type: string
      responses:
        '200':
          description: A page of transfer history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferHistoryPage'
        '400':
          description: Malformed cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: The caller has no account yet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /deposits:
    post:
      operationId: createDeposit
      summary: >
        Start a Stripe-funded top-up: creates a pending deposit and a
        matching Stripe PaymentIntent, returning its client_secret for
        Stripe.js/Elements to confirm directly with Stripe. account_id is
        never taken from the request body — it's derived from the bearer
        token. A successful client-side confirmPayment does NOT mean the
        deposit is credited yet — see GET /deposits/{id}.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepositRequest'
      responses:
        '201':
          description: Pending deposit created and a Stripe PaymentIntent started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDepositResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: The caller has no account yet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Account is not active.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /deposits/{id}:
    get:
      operationId: getDeposit
      summary: >
        Poll a deposit's status. 'succeeded' means Stripe confirmed the
        card charge; it does NOT mean the balance moved yet. 'credited'
        is the only status that means the ledger actually reflects it —
        the frontend must poll until 'credited' (or 'failed'), never stop
        at 'succeeded'.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The deposit's current state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deposit'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: >
            No such deposit, or it belongs to a different account —
            deliberately indistinguishable, so this endpoint can never be
            used to enumerate another account's deposit ids.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Malformed request body or invalid field value.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing, invalid, or expired bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: >
            A stable, machine-readable snake_case code (e.g.
            "insufficient_funds", "invalid_iban", "recipient_not_found") —
            never a human-readable sentence. The backend deliberately
            doesn't own display copy or its language; translating a code
            into a message for the signed-in user's locale is the
            frontend's job.
        attempts_remaining:
          type: integer
          description: >
            Only present on /auth/verify-email wrong-code responses — the
            number of guesses left before the code is locked out.
      example:
        error: invalid_credentials
...
    CreateTransferRequest:
      type: object
      required: [recipient_iban, amount]
      properties:
        recipient_iban:
          type: string
          description: >
            The recipient's IBAN — this system's one user-facing account
            identifier (see AccountWithBalance.iban). Only IBANs for this
            bank's own code are resolvable; a structurally valid IBAN for
            another institution is rejected (400), not silently treated as
            "not found" (also 400, but a distinct message) — see README,
            "Honest limitations".
          example: IE34ZZZZ00004234567890
        amount:
          type: integer
          format: int64
          description: Minor units (cents), must be positive.
    Transfer:
      type: object
      required:
        - id
        - idempotency_key
        - sender_account_id
        - recipient_account_id
        - amount
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        idempotency_key:
          type: string
        sender_account_id:
          type: string
          format: uuid
        recipient_account_id:
          type: string
          format: uuid
        amount:
          type: integer
          format: int64
        status:
          type: string
          enum: [pending, completed, failed, rejected]
        failure_reason:
          type: string
          description: >
            Two disjoint vocabularies depending on status: a ledger failure
            code when status is "failed" (e.g. "insufficient_funds",
            "account_not_found", "invalid_amount", "ledger_internal_error"),
            or fraud-svc's triggered rule name when status is "rejected"
            ("amount_threshold", "velocity_count", "velocity_sum"). The two
            statuses are mutually exclusive, so which vocabulary applies is
            never ambiguous.
        ledger_transaction_id:
          type: string
          format: uuid
          description: Only present when status is completed.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TransferResult:
      allOf:
        - $ref: '#/components/schemas/Transfer'
        - type: object
          properties:
            message:
              type: string
              description: >
                Only present when the outcome is uncertain (202) — e.g.
                "transfer status unknown, still processing".
    CreateDepositRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          format: int64
          description: >
            Minor units (cents). Bounded server-side to €0.50–€10,000.00
            (services/transfers-svc/deposit.go's depositMinAmount/
            depositMaxAmount).
    CreateDepositResponse:
      type: object
      required: [deposit_id, client_secret]
      properties:
        deposit_id:
          type: string
          format: uuid
        client_secret:
          type: string
          description: >
            Stripe PaymentIntent client_secret — hand this to
            Stripe.js/Elements to confirm the payment directly with
            Stripe. Never reuse across a retry after a decline: a new
            attempt means a new POST /deposits call and a new
            client_secret.
    Deposit:
      type: object
      required:
        - id
        - account_id
        - amount
        - currency
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        amount:
          type: integer
          format: int64
        currency:
          type: string
          example: eur
        status:
          type: string
          enum: [pending, succeeded, credited, failed, refunded]
          description: >
            'succeeded' means Stripe confirmed the card charge only.
            'credited' means the ledger balance actually reflects it —
            the status pollers must wait for. 'refunded' is only reachable
            from 'credited' (a charge.refunded event after this service
            already credited it); a poller that hasn't already stopped at
            'credited' should treat 'refunded' as terminal too.
        stripe_payment_intent_id:
          type: string
        ledger_transaction_id:
          type: string
          format: uuid
        failure_reason:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TransferHistoryPage:
      type: object
      required: [transfers]
      description: >
        Despite the name (kept for compatibility with the /transfers path
        and this field's existing key), this is the caller's UNIFIED
        operations feed — transfers, deposits, and withdrawals
        interleaved newest-first — not transfers alone. See
        services/transfers-svc/history.go's historyEntry doc comment.
      properties:
        transfers:
          type: array
          items:
            $ref: '#/components/schemas/TransferHistoryEntry'
        next_cursor:
          type: string
          description: >
            Pass as ?cursor= on the next request. Absent when there are no
            more pages.
    TransferHistoryEntry:
      type: object
      required: [type, id, amount, status, created_at, updated_at]
      properties:
        type:
          type: string
          enum: [transfer, deposit, withdrawal]
        id:
          type: string
          format: uuid
        amount:
          type: integer
          format: int64
          description: Minor units (cents).
        status:
          type: string
          description: >
            Vocabulary depends on type: transfer is one of [pending,
            completed, failed, rejected]; deposit is one of [pending,
            succeeded, credited, failed, refunded]; withdrawal is one of
            [payout_simulated, failed].
        failure_reason:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        direction:
          type: string
          enum: [outgoing, incoming]
          description: transfer only — absent for deposit/withdrawal entries.
        counterparty_iban:
          type: string
          description: transfer only — absent for deposit/withdrawal entries.